Core: Remove Vector3 and WoWPoint and replace with System.Numerics.Vector3 instead

+++ Core/Addon/ISquareReader.cs
@@ -2,7 +2,7 @@
 {
     public interface ISquareReader
     {
-        double GetFixedPointAtCell(int indexl);
+        float GetFixedPointAtCell(int indexl);
 
         int GetIntAtCell(int index);
